Wired services in order to check for existent objects

+++ src/main/java/main/lab1/exceptions/TaskNotFoundException.java
@@ -1,7 +1,7 @@
 package main.lab1.exceptions;
 
 public class TaskNotFoundException extends RuntimeException {
-    public TaskNotFoundException(int id) {
+    public TaskNotFoundException(long id) {
         super("Task with id = " + id + " not found.");
     }
 }
